move referendum util to common package
