+++ packages/next-common/utils/referendumUtil.js
@@ -1,5 +1,11 @@
 import BigNumber from "bignumber.js";
-import { Conviction, convictionToLockX, getConviction, isAye } from "next-common/utils/referendumCommon";
+import compareRationals from "next-common/utils/democracy/rational";
+import {
+  Conviction,
+  convictionToLockX,
+  getConviction,
+  isAye,
+} from "next-common/utils/referendumCommon";
 
 export { isAye, Conviction, getConviction, convictionToLockX };
 
@@ -100,4 +106,34 @@ export async function getAddressVote(api, referendumIndex, address) {
   }
 
   return null;
+}
+
+export function calcPassing(referendumInfo, totalIssuance) {
+  if (!referendumInfo) {
+    return false;
+  }
+
+  const ayes = new BigNumber(referendumInfo.tally.ayes);
+  const nays = new BigNumber(referendumInfo.tally.nays);
+  const threshold = referendumInfo.threshold;
+
+  if (threshold.toLowerCase() === "simplemajority") {
+    return ayes.gt(nays);
+  }
+
+  const turnout = new BigNumber(referendumInfo.tally.turnout);
+  const sqrtTurnout = turnout.sqrt();
+  const sqrtElectorate = BigNumber.max(totalIssuance, turnout).sqrt();
+
+  if (sqrtTurnout.isZero() || sqrtElectorate.isZero()) {
+    return false;
+  }
+
+  if (threshold.toLowerCase() === "supermajorityapprove") {
+    return compareRationals(nays, sqrtTurnout, ayes, sqrtElectorate);
+  } else if (threshold.toLowerCase() === "supermajorityagainst") {
+    return compareRationals(nays, sqrtElectorate, ayes, sqrtTurnout);
+  }
+
+  return false;
 }
